move Planck temperature generation to PhysicsUtil
